Added "Assign Payment Method" event.

+++ src/main/com/topcoder/web/ejb/pacts/BasePayment.java
@@ -103,7 +103,7 @@ public abstract class BasePayment implements Constants, java.io.Serializable {
     private boolean charity = false;
     private double totalAmount;
     private int installmentNumber;
-    private int methodId = Constants.NOT_SET_PAYMENT_METHOD_ID;
+    private long methodId = Constants.NOT_SET_PAYMENT_METHOD_ID;
 
     // If positive, the payment is for that contract, so a row relating the contact and the payment will be inserted
     private long contractId = 0;
@@ -628,11 +628,11 @@ public abstract class BasePayment implements Constants, java.io.Serializable {
         this.totalAmount = totalAmount;
     }
 
-    public int getMethodId() {
+    public long getMethodId() {
         return methodId;
     }
 
-    public void setMethodId(int methodId) {
+    public void setMethodId(long methodId) {
         this.methodId = methodId;
     }
